+++ lib/filter_list.dart
@@ -1,3 +1,4 @@
 library filter_list;
 
 export 'src/filter_list_dialog.dart';
+export 'src/theme/theme.dart';