+++ lib/src/theme/theme.dart
@@ -1,0 +1,2 @@
+export 'filter_list_theme.dart';
+export 'choice_chip_theme.dart';